requirements.txt
alembic==1.13.2
asgiref==3.8.1
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
distlib==0.3.8
distro==1.7.0
Django==5.1.1
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.1
greenlet==3.1.1
gunicorn==22.0.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.5
MarkupSafe==2.1.5
mock==5.1.0
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.8.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.4.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uWSGI==2.0.27
virtualenv==20.26.2
wadllib==1.3.9
zipp==1.0.0
